[18] Add License (#20)
* Add LICENSE file

We decided to use the MIT license.

Note: it's fine to use 2022 as the year. That's the year when we added
the license and it's valid from there on (does not need to be renewed
each year).

* Update package.json info

Add author and repository information to the package.json file. This
helps people to find the package on npm.