[#25] Add support for expand nodes
For now, ADF `expand` and `nestedExpand` nodes will be skipped, meaning
the nodes themselves do not create a counterpart in the resulting MDAST
tree. Their child nodes however will be processed and attached to the
output tree.

An alternative to skipping "expand" nodes could be to represent them as
a sequence of MDAST nodes:

1. HTML node: `<details>`
2. HTML node: `<summary>` + expand node title + `</summary>`
3. MDAST nodes for the "expand" node content (the current output)
4. HTML node: `</details>`

This produces adequate output when converting the ADF to Markdown and
HTML. However, it also comes with its own set of challenges and arguably
makes the MDAST tree representation look rather odd, with HTML nodes
containing invalid HTML content. The opening and closing `<details>`
tags are not valid on their own – though the MDAST documentation
explicitly states that HTML nodes do not need to contain "valid or
complete HTML" (https://github.com/syntax-tree/mdast#html).

For the moment, we go with the "simple" path of ignoring expand nodes.

https://github.com/bitcrowd/mdast-util-from-adf/issues/25